refactor(kanban): extract models + schema into kanban_db_models (kanban_db.py god-file slice R1)
